Removed a couple sign errors in a hacky way. Should really size_t that.

+++ src/Luna/LunaFile.cpp
@@ -13,7 +13,7 @@
   1. The origin of this software must not be misrepresented; you must not
      claim that you wrote the original software. If you use this software
      in a product, an acknowledgment in the product documentation would be
-     appreciated but is not required. 
+     appreciated but is not required.
   2. Altered source versions must be plainly marked as such, and must not be
      misrepresented as being the original software.
   3. This notice may not be removed or altered from any source distribution.
@@ -185,7 +185,7 @@ static LunaArchive *LunaArchive_Open(Uint32 priority,
         
         entrySize = sizeof(ArchiveEntry) * header.entrycount;
         
-        if (PL_File_GetSize(fileHandle) < (entrySize + sizeof(ArchiveHeader))) {
+        if (PL_File_GetSize(fileHandle) < (int64_t)(entrySize + sizeof(ArchiveHeader))) {
             break;
         }
