+++ src/Luna/LunaFontSprite.cpp
@@ -13,7 +13,7 @@
   1. The origin of this software must not be misrepresented; you must not
      claim that you wrote the original software. If you use this software
      in a product, an acknowledgment in the product documentation would be
-     appreciated but is not required. 
+     appreciated but is not required.
   2. Altered source versions must be plainly marked as such, and must not be
      misrepresented as being the original software.
   3. This notice may not be removed or altered from any source distribution.
@@ -124,7 +124,7 @@ LFONTSPRITE LunaFontSprite::CreateFromFile(const char *pFileName,
         }
         
         fontBaseSize = sizeof(LFDHeader) + (sizeof(Uint16) * CODE_TABLE_SIZE) + (32 * lfdHeader->sheetCount);
-        if (fileSize < (fontBaseSize + (sizeof(LFDCharEntry) * lfdHeader->fontMax))) {
+        if (fileSize < (int64_t)(fontBaseSize + (sizeof(LFDCharEntry) * lfdHeader->fontMax))) {
             break;
         }
